Created a reusable `RefreshableFlow`
